+++ CHANGELOG.md
@@ -5,9 +5,10 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 
 ## [Unreleased]
 ### Added
-- add [`no-mutable-exports`] rule ([#290], thanks [@josh])
+- [`newline-after-import`], new rule. ([#245], thanks [@singles])
+- [`no-mutable-exports`], new rule. ([#290], thanks [@josh])
 
-## resolvers/webpack: Unreleased
+## resolvers/webpack/0.2.4 - 2016-04-29
 ### Changed
 - automatically find webpack config with `interpret`-able extensions ([#287], thanks [@taion])
 
@@ -187,6 +188,7 @@ for info on changes for earlier releases.
 [`no-nodejs-modules`]: ./docs/rules/no-nodejs-modules.md
 [`order`]: ./docs/rules/order.md
 [`named`]: ./docs/rules/named.md
+[`newline-after-import`]: ./docs/rules/newline-after-import.md
 
 [#289]: https://github.com/benmosher/eslint-plugin-import/pull/289
 [#288]: https://github.com/benmosher/eslint-plugin-import/pull/288
@@ -198,6 +200,7 @@ for info on changes for earlier releases.
 [#250]: https://github.com/benmosher/eslint-plugin-import/pull/250
 [#247]: https://github.com/benmosher/eslint-plugin-import/pull/247
 [#243]: https://github.com/benmosher/eslint-plugin-import/pull/243
+[#245]: https://github.com/benmosher/eslint-plugin-import/pull/245
 [#241]: https://github.com/benmosher/eslint-plugin-import/pull/241
 [#239]: https://github.com/benmosher/eslint-plugin-import/pull/239
 [#228]: https://github.com/benmosher/eslint-plugin-import/pull/228